requirements.txt
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.0
colorama==0.4.4
Flask==2.0.0
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
idna==2.10
itsdangerous==2.0.0
Jinja2==3.0.0
MarkupSafe==2.0.0
pycparser==2.20
pycryptodome==3.10.1
pymongo==3.11.3
requests==2.25.1
selenium==3.141.0
six==1.16.0
urllib3==1.26.4
Werkzeug==2.0.0
